refactor: hooks 폴더를 서버 상태와 클라이언트 상태를 기준으로 명확히 구분
