+++ src/apis/follow/deleteFollow.ts
@@ -1,15 +1,12 @@
 import { isAxiosError } from 'axios';
 
-import { ARTWORK, USER } from '@/constants/API';
+import { USER } from '@/constants/API';
 import {
   COMMON_ERROR_MESSAGE,
   FOLLOW_ERROR_MESSAGE,
 } from '@/constants/errorMessage';
 
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-
 import { authorizedClient } from '..';
-import postFollow from './postFollow';
 
 const deleteFollow = async (userId: number) => {
   try {
@@ -40,30 +37,4 @@ const deleteFollow = async (userId: number) => {
   }
 };
 
-const useToggleFollow = (
-  isFollowing: boolean,
-  setIsFollowing: (value: boolean) => void,
-) => {
-  const queryClient = useQueryClient();
-
-  return useMutation<void, Error, number>({
-    mutationFn: async (userId: number) => {
-      isFollowing === true
-        ? await deleteFollow(userId)
-        : await postFollow(userId);
-    },
-
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: [ARTWORK.followedArtists],
-      });
-      setIsFollowing(!isFollowing);
-    },
-
-    onError: (error) => {
-      console.error('팔로우 상태 변경 에러: ', error.message);
-    },
-  });
-};
-
-export default useToggleFollow;
+export default deleteFollow;
